adding preliminary locale selection/settings support
This commit adds a basic link on the splash page to allow users to
select their desired language. This takes them to the settings page,
which currently only offers configuration of their desired locale. Still
to do: addition of a settings icon at the top of the splash/stories
screen.